Artist Manifesto with html link for first draft Twine
The purpose of this project is to put the player’s minds through a zombie apocalypse, as if they were trying to survive through a catastrophe of that nature. I want to simulate a situation like this, a survival in a world that is basically gone. The project accomplishes many different options to choose from, as I want it to be more of an open world place that keeps the user on their toes, creating outcomes that they were not expecting, or just given choices that the player would not have thought of for in this game. I hope to achieve a mood of suspense, and dramatic events in addition with the game being thought-provoking. I was the user to have interactivity and to base decisions off what they think is correct, and I want the game to reciprocate the interactivity back. I hope to achieve agency by giving multiple options to choose from, and immersion by keeping the player focused throughout the entirety of the story. I will try to reach those by analyzing what the outcome should be from situations, and to add some type of humor to keep the attention span of the player. Also, I would like to add multiple different features from Twine, and to play with the style, text, fonts, and other pieces in relation to the game.